+++ examples/kotlin/build.gradle.kts
@@ -22,6 +22,7 @@ dependencies {
   implementation(projects.xefSql)
   implementation(projects.xefTokenizer)
   implementation(projects.xefGpt4all)
+  implementation(projects.xefOpenai)
   implementation(libs.kotlinx.serialization.json)
   implementation(libs.logback)
   implementation(libs.klogging)
