+++ java/build.gradle.kts
@@ -8,6 +8,7 @@ plugins {
 
 dependencies {
     implementation(projects.xefCore)
+    implementation(projects.xefOpenai)
     implementation(projects.xefPdf)
     implementation("com.fasterxml.jackson.core:jackson-databind:2.15.2")
     implementation("com.fasterxml.jackson.module:jackson-module-jsonSchema:2.15.2")
