+++ kotlin/build.gradle.kts
@@ -39,6 +39,7 @@ kotlin {
     val commonMain by getting {
       dependencies {
           api(projects.xefCore)
+          api(projects.xefOpenai)
       }
     }
   }
