+++ scala/src/main/scala/com/xebia/functional/xef/scala/auto/package.scala
@@ -9,11 +9,12 @@ import io.circe.Decoder
 import io.circe.parser.parse
 import com.xebia.functional.xef.auto.AIKt
 import com.xebia.functional.xef.auto.AIRuntime
+import com.xebia.functional.xef.auto.llm.openai.OpenAIRuntime
 import com.xebia.functional.xef.auto.serialization.JsonSchemaKt
 import com.xebia.functional.xef.pdf.PDFLoaderKt
 import com.xebia.functional.tokenizer.ModelType
 import com.xebia.functional.xef.llm._
-import com.xebia.functional.xef.llm.openai._
+import com.xebia.functional.xef.auto.llm.openai._
 import com.xebia.functional.xef.scala.textsplitters.TextSplitter
 import com.xebia.functional.xef.llm.models.images.*
 
@@ -26,7 +27,7 @@ type AI[A] = AIScope ?=> A
 def ai[A](block: AI[A]): A =
   LoomAdapter.apply { cont =>
     AIKt.AIScope[A](
-      AIRuntime.defaults[A](),
+      OpenAIRuntime.defaults[A](),
       { (coreAIScope, _) =>
         given AIScope = AIScope.fromCore(coreAIScope)
